+++ src/main/kotlin/cloud/glitchdev/rfu/config/categories/DropsSettings.kt
@@ -1,6 +1,7 @@
 package cloud.glitchdev.rfu.config.categories
 
 import cloud.glitchdev.rfu.config.Category
+import cloud.glitchdev.rfu.constants.fishing.RareDropDisplayDataType
 import cloud.glitchdev.rfu.constants.skyblock.Dyes
 import cloud.glitchdev.rfu.constants.fishing.RareDrops
 import cloud.glitchdev.rfu.feature.drops.RareDropAlert
@@ -94,5 +95,41 @@ object DropsSettings : Category("Drops") {
             "Preview Title",
             "Shows a preview of the rare drop title on screen."
         ) { rareDropTitleAlert }
+    }
+
+    init {
+        dualSeparator {
+            title = "Rare Drops Display"
+            description = "Track your rare drops!"
+        }
+    }
+
+    var rareDropsDisplay by reloadableBoolean(true) {
+        name = Literal("Toggle")
+        description = Literal("Enables the Rare Drops display")
+    }
+
+    var rareDropsOnlyWhenFishing by boolean(true) {
+        name = Literal("Only display when fishing")
+        description = Literal("Only show the rare drops when fishing")
+        condition = { rareDropsDisplay }
+    }
+
+    var rareDropsDisplayDyes by boolean(true) {
+        name = Literal("Include Dyes")
+        description = Literal("Include Dyes in the display")
+        condition = { rareDropsDisplay }
+    }
+
+    var displayedRareDrops by enums(*RareDrops.entries.toTypedArray()) {
+        name = Literal("Displayed Rare Drops")
+        description = Literal("Select which rare drops will appear on the display")
+        condition = { rareDropsDisplay }
+    }
+
+    var rareDropsDisplayDataOrder by draggable(*RareDropDisplayDataType.entries.toTypedArray()) {
+        name = Literal("Display Data Order")
+        description = Literal("Drag to reorder the data shown for each drop.")
+        condition = { rareDropsDisplay }
     }
 }